Added room layout utility
