+++ Runtime/Internal/RoomLayout/RoomLayoutUtil.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ed9e60d8645c814387dc354d4531bc6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
